moved tokens components and buttons into their own folders
