+++ ngVioletMarketplaceMaterial/src/app/modules/shared/shared.module.ts
@@ -9,8 +9,8 @@ import { AboutUsComponent } from './components/about-us/about-us.component';
 import { NavbarComponent } from './components/navbar/navbar.component';
 import { SharedRoutingModule } from './shared-routing.module';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { DoubleGradButtonComponent } from './components/double-grad-button/double-grad-button.component';
-import { NavGradButtonComponent } from './components/nav-grad-button/nav-grad-button.component';
+import { DoubleGradButtonComponent } from './components/buttons/double-grad-button/double-grad-button.component';
+import { NavGradButtonComponent } from './components/buttons/nav-grad-button/nav-grad-button.component';
 import { UserRoutingModule } from '../user/user-routing.module';
 import { AngularSvgIconModule } from 'angular-svg-icon';
 import { FormFieldComponent } from './components/forms/form-field/form-field.component';
