+++ ngVioletMarketplaceMaterial/src/app/modules/tokens/components/big-token/big-token-card/big-token-card.component.spec.ts
