+++ ngVioletMarketplaceMaterial/src/app/modules/tokens/components/big-token/big-token-list/big-token-list.component.scss
