+++ ngVioletMarketplaceMaterial/src/app/modules/tokens/components/small-token/token-card/token-card.component.html
