+++ ngVioletMarketplaceMaterial/src/app/modules/tokens/components/small-token/token-list/token-list.component.scss
