+++ ngVioletMarketplaceMaterial/src/app/modules/tokens/components/token-detail/token.component.ts
