+++ ngVioletMarketplaceMaterial/src/app/modules/tokens/token-routing.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { TokenComponent } from './components/token/token.component';
+import { TokenComponent } from './components/token-detail/token.component';
+import { TokenListComponent } from './components/small-token/token-list/token-list.component';
 
 const routes: Routes = [
   {
@@ -10,6 +11,9 @@ const routes: Routes = [
   {
     path: 'token/id/:id',
     component: TokenComponent,
+  }, {
+    path: 'tokens-list',
+    component: TokenListComponent,
   },
 ];
